+++ CHANGELOG.md
@@ -25,6 +25,7 @@ avoid adding features or APIs which do not map onto the
 ### Fixed
 - Fix build for PostgreSQL 12 (#4, thanks @Komzpa)
 ### Removed
+- Removed H3 core library version check, since we know which version we are linking.
 - Removed `h3_hex_area_m2`, `h3_hex_area_km2`, `h3_edge_length_m` and `h3_edge_length_km` in favor of `h3_hex_area` and `h3_edge_length`.
 - Removed `hex_range`, `hex_ranges` and `hex_range_distances` to align with `binding-functions`.