+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Because H3-pg is versioned in lockstep with the H3 core library, please
 avoid adding features or APIs which do not map onto the
 [H3 core API](https://uber.github.io/h3/#/documentation/api-reference/).
 
-## [Unreleased]
+## [3.5.0] - 2019-08-01
 ### Added
 - `h3_get_faces` function to find icosahedron faces for an index.
 - `h3_hex_area` replacing `h3_hex_area_m2` and `h3_hex_area_km2`.
@@ -95,7 +95,8 @@ avoid adding features or APIs which do not map onto the
 ### Added
 - First public release.
 
-[Unreleased]: https://github.com/bytesandbrains/h3-pg/compare/v3.4.1...HEAD
+[Unreleased]: https://github.com/bytesandbrains/h3-pg/compare/v3.5.0...HEAD
+[3.5.0]: https://github.com/bytesandbrains/h3-pg/compare/v3.4.1...v.3.5.0
 [3.4.1]: https://github.com/bytesandbrains/h3-pg/compare/v3.4.0...v3.4.1
 [3.4.0]: https://github.com/bytesandbrains/h3-pg/compare/v1.0.6...v3.4.0
 [1.0.6]: https://github.com/bytesandbrains/h3-pg/compare/v1.0.5...v1.0.6